% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rle_add}
\alias{rle_add}
\title{Add two rle objects, element-wise}
\usage{
rle_add(rle_a, rle_b)
}
\arguments{
\item{rle_a}{An R rle list object}

\item{rle_b}{An R rle list object}
}
\value{
An R rle list object containing the sum
}
\description{
Add two rle objects, element-wise
}
\examples{
rle_a <- list(lengths = c(2, 3), values = c(10, 20))
rle_b <- list(lengths = c(3, 2), values = c(5, 15))
result <- rle_add(rle_a, rle_b)

stopifnot(all.equal(result$lengths, c(2L, 1L, 2L)))
stopifnot(all.equal(result$values, c(15, 25, 35)))
}
