% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rle_compact}
\alias{rle_compact}
\title{Compact an rle object by merging identical runs}
\usage{
rle_compact(rle)
}
\arguments{
\item{rle}{An R rle list object}
}
\value{
An R rle list object with identical runs merged
}
\description{
Compact an rle object by merging identical runs
}
\examples{
uncompacted <- list(lengths = c(2, 3, 1), values = c(5, 5, 10))
res1 <- rle_compact(uncompacted)
stopifnot(all.equal(res1, list(lengths = c(5, 1), values = c(5, 10))))

compacted <- list(lengths = c(5, 2), values = c(1, 2))
stopifnot(all.equal(rle_compact(compacted), compacted))

nas <- list(lengths = c(1, 2), values = c(NA_real_, NA_real_))
res2 <- rle_compact(nas)
stopifnot(all.equal(res2, list(lengths = 3, values = NA_real_)))
}
