% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rle_cummax}
\alias{rle_cummax}
\title{Cumulative Max of an rle object}
\usage{
rle_cummax(rle)
}
\arguments{
\item{rle}{An R rle list object}
}
\value{
An R rle list object
}
\description{
Cumulative Max of an rle object
}
