% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rle_eq}
\alias{rle_eq}
\title{Returns true if both rle object have the same elements}
\usage{
rle_eq(rle_a, rle_b)
}
\arguments{
\item{rle_a}{An R rle list object}

\item{rle_b}{An R rle list object}
}
\value{
An R rle list object containing 1.0 (TRUE), 0.0 (FALSE) or NaN
}
\description{
Returns true if both rle object have the same elements
}
