% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rle_filter}
\alias{rle_filter}
\title{Filter an rle vector by another logical rle vector}
\usage{
rle_filter(rle_data, rle_mask)
}
\arguments{
\item{rle_data}{An R rle list object with data to filter}

\item{rle_mask}{An R rle list object with logical mask}
}
\value{
A filtered rle list object
}
\description{
Filter an rle vector by another logical rle vector
}
