% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rle_length}
\alias{rle_length}
\title{Total length of the rle object (number of elements before encoding)}
\usage{
rle_length(rle)
}
\arguments{
\item{rle}{An R rle list object}
}
\value{
Total length
}
\description{
Total length of the rle object (number of elements before encoding)
}
\examples{
rle_a <- list(lengths = c(2, 3, 1), values = c(1, 2, 3))
stopifnot(rle_length(rle_a) == 6)

empty <- list(lengths = integer(0), values = numeric(0))
stopifnot(rle_length(empty) == 0)
}
