% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rle_mean}
\alias{rle_mean}
\title{Mean of all values in an rle object}
\usage{
rle_mean(rle, na_rm = FALSE)
}
\arguments{
\item{rle}{An R rle list object}

\item{na_rm}{If TRUE, NA values are ignored. Default FALSE.}
}
\value{
A numeric scalar
}
\description{
Mean of all values in an rle object
}
\examples{
rle_a <- list(lengths = c(2, 2), values = c(0.0, 4.0))
stopifnot(rle_mean(rle_a) == 2.0)

rle_na <- list(lengths = c(3, 1), values = c(6.0, NA_real_))
stopifnot(is.na(rle_mean(rle_na)))
stopifnot(rle_mean(rle_na, na_rm = TRUE) == 6.0)
}
