% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rle_slice}
\alias{rle_slice}
\title{Extract a slice of an rle vector}
\usage{
rle_slice(rle, start, end)
}
\arguments{
\item{rle}{An R rle list object}

\item{start}{The 1-based starting index}

\item{end}{The 1-based ending index}
}
\value{
A sliced R rle list object
}
\description{
Extract a slice of an rle vector
}
