% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rle_sum}
\alias{rle_sum}
\title{Sum of all values in an rle object}
\usage{
rle_sum(rle, na_rm = FALSE)
}
\arguments{
\item{rle}{An R rle list object}

\item{na_rm}{If TRUE, NA values are ignored.}
}
\value{
Sum of values, or NA if any value is NA and na_rm is FALSE
}
\description{
Sum of all values in an rle object
}
\examples{
rle_a <- list(lengths = c(2, 3), values = c(4.0, 2.0))
stopifnot(rle_sum(rle_a) == 14.0)

rle_na <- list(lengths = c(2, 1), values = c(3.0, NA_real_))
stopifnot(is.na(rle_sum(rle_na)))
stopifnot(rle_sum(rle_na, na_rm = TRUE) == 6.0)
}
