% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rle_value_counts}
\alias{rle_value_counts}
\title{Unique values and their frequencies of an RLE object}
\usage{
rle_value_counts(rle)
}
\arguments{
\item{rle}{An R rle list object}
}
\value{
A list containing unique values and their total counts
}
\description{
Unique values and their frequencies of an RLE object
}
