% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rle_which}
\alias{rle_which}
\title{Find indices of TRUE (1.0) values in an RLE object}
\usage{
rle_which(rle)
}
\arguments{
\item{rle}{An R rle list object (where 1.0 represents TRUE)}
}
\value{
An integer vector of 1-based indices
}
\description{
Find indices of TRUE (1.0) values in an RLE object
}
