% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rle_add}
\alias{rle_add}
\title{Add two rle objects, element-wise}
\usage{
rle_add(rle_a, rle_b)
}
\arguments{
\item{rle_a}{An R rle list object}

\item{rle_b}{An R rle list object}
}
\value{
An R rle list object containing the sum
}
\description{
Add two rle objects, element-wise
}
