% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rle_compact}
\alias{rle_compact}
\title{Compact an RLE object by merging identical runs}
\usage{
rle_compact(rle)
}
\arguments{
\item{rle}{An R rle list object}
}
\value{
An R rle list object with identical runs merged
}
\description{
Compact an RLE object by merging identical runs
}
