% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rle_cummin}
\alias{rle_cummin}
\title{Cumulative Min of an rle object}
\usage{
rle_cummin(rle)
}
\arguments{
\item{rle}{An R rle list object}
}
\value{
An R rle list object
}
\description{
Cumulative Min of an rle object
}
