% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rle_slice}
\alias{rle_slice}
\title{Extract a slice of an rle vector}
\usage{
rle_slice(rle, start, end)
}
\arguments{
\item{rle}{An R rle list object}

\item{start}{The 1-based starting index}

\item{end}{The 1-based ending index}
}
\value{
A sliced R rle list object
}
\description{
Extract a slice of an rle vector
}
\examples{
rle_in <- list(lengths = c(3, 3, 3), values = c(10, 20, 30))

res1 <- rle_slice(rle_in, 2, 7)
stopifnot(isTRUE(all.equal(res1, list(lengths = c(2, 3, 1), values = c(10, 20, 30)))))

res2 <- rle_slice(rle_in, 1, 9)
stopifnot(all.equal(res2, rle_in))

res3 <- rle_slice(rle_in, 5, 5)
stopifnot(all.equal(res3, list(lengths = 1, values = 20)))
res4 <- rle_slice(rle_in, 8, 20)
stopifnot(all.equal(res4, list(lengths = 2, values = 30)))
}
