% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rle_value_counts}
\alias{rle_value_counts}
\title{Unique values and their frequencies of an rle object}
\usage{
rle_value_counts(rle)
}
\arguments{
\item{rle}{An R rle list object}
}
\value{
A list containing unique values and their total counts
}
\description{
Unique values and their frequencies of an rle object
}
\examples{
rle_in <- list(lengths = c(2, 3, 4), values = c(10, 20, 10))
res1 <- rle_value_counts(rle_in)
# No guarantee on the order in output
stopifnot(all.equal(res1$counts[res1$values == 10], 6))
stopifnot(all.equal(res1$counts[res1$values == 20], 3))

empty <- list(lengths = integer(0), values = numeric(0))
res3 <- rle_value_counts(empty)
stopifnot(all.equal(length(res3$values), 0))
}
